requirements.txt
amqp==5.3.1
anyio==4.11.0
asgiref==3.9.1
async-timeout==5.0.1
attrs==25.3.0
azure-storage-blob==12.27.0
billiard==4.2.2
black==25.1.0
Brotli==1.1.0
celery==5.5.3
celery-types==0.23.0
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
coverage==7.12.0
cryptography==45.0.6
dj-database-url==3.0.1
dj-rest-auth==7.0.1
Django==5.2.5
django-allauth==65.11.1
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-otp==1.6.1
django-phonenumber-field==8.3.0
django-phonenumbers==1.0.1
django-storages==1.14.6
django-stubs==5.2.2
django-stubs-ext==5.2.2
djangorestframework==3.16.1
djangorestframework-stubs==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.28.0
exceptiongroup==1.3.0
Faker==38.2.0
fido2==2.0.0
fonttools==4.60.1
gevent==25.9.1
greenlet==3.2.4
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
isodate==0.7.2
jsonschema==4.25.0
jsonschema-specifications==2025.4.1
kombu==5.5.4
mypy==1.18.2
mypy_extensions==1.1.0
numpy==2.2.6
opencv-python-headless==4.12.0.88
packaging==25.0
pathspec==0.12.1
phonenumbers==9.0.16
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
pycparser==2.22
PyJWT==2.10.1
pyotp==2.9.0
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-decouple==3.8
pytokens==0.1.10
PyYAML==6.0.2
qrcode==8.2
redis==7.0.1
referencing==0.36.2
reportlab==4.4.4
requests==2.32.4
rpds-py==0.27.0
setuptools==80.9.0
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
tomli==2.2.1
types-defusedxml==0.7.0.20250822
types-django-filter==25.2.0.20251010
types-docutils==0.22.2.20251006
types-Markdown==3.10.0.20251106
types-psycopg2==2.9.21.20251012
types-Pygments==2.19.0.20250809
types-PySocks==1.7.1.20251001
types-python-dateutil==2.9.0.20251115
types-PyYAML==6.0.12.20250516
types-qrcode==8.2.0.20250914
types-reportlab==4.4.4.20250926
types-requests==2.32.4.20250611
typing_extensions==4.14.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
watchfiles==0.19.0
wcwidth==0.2.14
webencodings==0.5.1
zope.event==6.0
zope.interface==8.0.1
zopfli==0.2.3.post1
